requirements.txt
antlr4-python3-runtime==4.13.0
pygls==1.3.1